section.cc 内部に重大な脆弱性があることに気付き，これを修正した． また，メモリーリークの危険がある部分の修正を行った． 主な変更点は以下の通り:
 - 各クラスのデストラクタに，delete を追加した．
 - Section::getsh_name() をパーサの内部で呼び出すように変更した．

シンボルテーブルに対する理解が深まった．
これまでシンボルテーブルとは，プログラム内部で使用されるあらゆるシンボルが一括で管理されているテーブルだと思っていたが，実際にはある一つのシンボルに対する「名前・値・スコープ・型」などをまとめた構造体であった．
したがって，シンボルテーブル構造体は，おそらくシンボルの数だけ存在する．
ダイナミックリンクに関するシンボルは.dynsym に格納され，それ以外は.symtab に格納されている．
実装の際には，セクションヘッダーテーブルと全く同じ要領で行えば良い．

+++ build.sh
@@ -1,6 +1,13 @@
 #!/bin/bash
 set -e
+if [ $# -ne 1 ]; then
+	echo -e "Please type option:\n"
+	echo -e "  -h : Elf header"
+	echo -e "  -S : Section header"
+	echo -e "  -p : Program header"
+	exit 1
+fi
 # g++ device.cc elf.cc readelf.cc
 g++ device.cc elf64.cc section64.cc program64.cc readelf.cc
-./a.out
+./a.out $1
 rm a.out
